requirements.txt
appdirs==1.4.3
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==20.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cycler==0.10.0
dbus-python==1.2.16
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
itsdangerous==1.1.0
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
matplotlib==3.3.4
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.20.1
oauthlib==3.1.0
opencv-python==4.5.1.48
pexpect==4.6.0
Pillow==8.1.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyGLM==1.99.3
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.5
pyOpenSSL==19.0.0
pyparsing==2.4.7
PyQt5==5.15.2
pyqt5-plugins==5.15.2.2.0.1
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
pyqt5-tools==5.15.2.3.0.2
pyrsistent==0.15.5
pyserial==3.4
PySide2==5.15.2
pytesseract==0.3.7
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-dotenv==0.15.0
PyYAML==5.3.1
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
shiboken2==5.15.2
simplejson==3.16.0
six==1.14.0
sos==4.0
ssh-import-id==5.10
systemd-python==234
Twisted==18.9.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
uWSGI==2.0.19.1
virtualenv==20.0.17
wadllib==1.3.3
Werkzeug==1.0.1
zipp==1.0.0
zope.interface==4.7.1
